// svelte/src/lib/types/userinfo.ts
export interface UserInfo {
    user_id?:        string;
    user_id_type?:   string;
    user_name:       string;
	password:        string
    first_name?:     string;
    last_name?:      string;
    email?:          string;
    user_address?:   string;
    phone_number?:   string;
    user_type:       string;
    redirect_url:    string;
    user_role:       string;
    user_status:     string;
    picture?:        string;
    locale?:         string;
    v_token?:        string;
}